base tests for ODE filter
